Move all converter/validator to individual java file
