+++ app/src/main/java/org/astraea/argument/converter/BooleanConverter.java
@@ -1,0 +1,10 @@
+package org.astraea.argument.converter;
+
+import com.beust.jcommander.IStringConverter;
+
+public class BooleanConverter implements IStringConverter<Boolean> {
+  @Override
+  public Boolean convert(String value) {
+    return Boolean.valueOf(value);
+  }
+}
